Reorganize utils and tests
