+++ src/errors/AuthError.ts
@@ -1,0 +1,5 @@
+export class AuthError extends Error {
+  constructor() {
+    super("Not authorized");
+  }
+}
