+++ src/errors/InvalidLoginError.ts
@@ -1,0 +1,5 @@
+export class InvalidLoginError extends Error {
+  constructor() {
+    super("Incorrect email or password");
+  }
+}
