+++ src/errors/MaxAttemptsError.ts
@@ -1,0 +1,5 @@
+export class MaxAttemptsError extends Error {
+  constructor() {
+    super("Too many attempts.");
+  }
+}
