+++ src/errors/MissingFieldError.ts
@@ -1,0 +1,5 @@
+export class MissingFieldError extends Error {
+  constructor(requiredField: string) {
+    super(`Must have ${requiredField}.`);
+  }
+}
