+++ src/errors/NotFoundError.ts
@@ -1,0 +1,5 @@
+export class NotFoundError extends Error {
+  constructor(recordType: string) {
+    super(`${recordType} not found.`);
+  }
+}
