+++ src/errors/WrongPasswordError.ts
@@ -1,0 +1,5 @@
+export class WrongPasswordError extends Error {
+  constructor() {
+    super("Invalid password");
+  }
+}
