+++ src/errors/index.ts
@@ -1,0 +1,6 @@
+export { AuthError } from './AuthError';
+export { InvalidLoginError } from './InvalidLoginError';
+export { MissingFieldError } from './MissingFieldError';
+export { NotFoundError } from './NotFoundError';
+export { MaxAttemptsError } from './MaxAttemptsError';
+export { WrongPasswordError } from './WrongPasswordError';
